Fix everything, in theory
There were two problems: The big'un and the lil'un. The big'un was that
the vertex and UV index buffers were made in the wrong scope, so all the
quads were storing the same buffers as they mutated. The lil'un was that
merged textures weren't being cached properly because we weren't, yanno,
*adding* them to the cache. This fixes both of those issues in one
über-cool commit!